% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ntd.R
\name{get_ntd}
\alias{get_ntd}
\title{Get NTD data}
\usage{
get_ntd(data_type = "adjusted", ntd_variable = "UPT", agency, modes = "MB")
}
\arguments{
\item{data_type}{Type of NTD data. Either "raw" for data released without adjustments or "adjusted" for data with adjustments.}

\item{ntd_variable}{Which variable to return. `UPT` for unlinked passenger trips, `VRM` for vehicle revenue miles, `VRH` for vehicle revenue hours, or `VOMS` for vehicles operated in maximum service.}

\item{agency}{Name of the transit agency to retrieve.}

\item{modes}{Transit mode to retrieve. Common modes include `MB` (bus), `CR` (commuter rail), `HR` (heavy rail), `LR` (light rail)}
}
\value{
A data frame of monthly NTD data with the requested ntd_variable in the `value` column
}
\description{
Get NTD data
}
\examples{
get_ntd(agency = "City of Madison", modes = c("MB", "DR"))
}
