% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ntd_url.R
\name{get_ntd_url}
\alias{get_ntd_url}
\title{Create NTD download URL}
\usage{
get_ntd_url(data_type = "adjusted")
}
\arguments{
\item{data_type}{Type of NTD data. Either "raw" for data released without adjustments or "adjusted" for data with adjustments.}
}
\value{
A string containing the URL to download the Excel file of the NTD data.
}
\description{
`get_ntd_url()` returns a URL needed to download to data from the NTD website
}
\examples{
get_ntd_url()
}
