% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntdr-package.R
\docType{package}
\name{ntdr-package}
\alias{ntdr}
\alias{ntdr-package}
\title{ntdr: Retrieve Data from the National Transit Database}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Downloads the latest 'National Transit Database' data, processes it, and returns in a tidy data format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vgxhc.github.io/ntdr/}
  \item \url{https://github.com/vgXhc/ntdr}
  \item Report bugs at \url{https://github.com/vgXhc/ntdr/issues}
}

}
\author{
\strong{Maintainer}: Harald Kliems \email{kliems@gmail.com} (\href{https://orcid.org/0000-0003-2542-0047}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [contributor]
}

}
\keyword{internal}
